feat: Update FlyingEye behavior and animations
- Changed FlyingEye death animation to not loop.
- Adjusted positions of various game objects in GameplayScene for better alignment.
- Added damageableDeath event to Damageable class to handle death events.
- Enhanced FadeRemoveBehaviour to include a fade delay before removal.
- Implemented FlyingEye movement and waypoint navigation logic.
- Updated project settings to reflect new build version 1.18.0-alpha.
- Created new build for version 1.18.0-alpha.
