--- Assets/FlyingEye.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9c5b7cab66182c24da83dd25457fcd64
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
