+++ Assets/Scenes/GameplayScene.unity
@@ -620,7 +620,7 @@ Transform:
   m_GameObject: {fileID: 812287389}
   serializedVersion: 2
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 2.73, y: 4, z: 0}
+  m_LocalPosition: {x: 2.73, y: 3, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
@@ -2694,7 +2694,7 @@ Transform:
   m_GameObject: {fileID: 2069201547}
   serializedVersion: 2
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 9, y: 4, z: 0}
+  m_LocalPosition: {x: 9, y: 3, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
@@ -2868,6 +2868,10 @@ PrefabInstance:
       propertyPath: m_Name
       value: HealthPickup
       objectReference: {fileID: 0}
+    - target: {fileID: 5806242851588284159, guid: 8a8ba9f7bfb23eb41addcd77c28aa74a, type: 3}
+      propertyPath: m_Layer
+      value: 12
+      objectReference: {fileID: 0}
     - target: {fileID: 6492907415954848632, guid: 8a8ba9f7bfb23eb41addcd77c28aa74a, type: 3}
       propertyPath: m_LocalPosition.x
       value: 3.65
@@ -2921,13 +2925,17 @@ PrefabInstance:
     serializedVersion: 3
     m_TransformParent: {fileID: 0}
     m_Modifications:
+    - target: {fileID: 3097971508342582203, guid: 525a057b766799d45a3905efa567a148, type: 3}
+      propertyPath: m_Layer
+      value: 10
+      objectReference: {fileID: 0}
     - target: {fileID: 5402913328238803636, guid: 525a057b766799d45a3905efa567a148, type: 3}
       propertyPath: m_LocalPosition.x
-      value: 8.77
+      value: 8.85
       objectReference: {fileID: 0}
     - target: {fileID: 5402913328238803636, guid: 525a057b766799d45a3905efa567a148, type: 3}
       propertyPath: m_LocalPosition.y
-      value: 2.13
+      value: 5.39
       objectReference: {fileID: 0}
     - target: {fileID: 5402913328238803636, guid: 525a057b766799d45a3905efa567a148, type: 3}
       propertyPath: m_LocalPosition.z
