+++ Assets/Scripts/Damageable.cs
@@ -6,7 +6,8 @@ using UnityEngine.Events;
 
 public class Damageable : MonoBehaviour
 {
-    public UnityEvent<int, Vector2> damageableHit; 
+    public UnityEvent<int, Vector2> damageableHit;
+    public UnityEvent damageableDeath;
     
     Animator animator;
 
@@ -68,6 +69,11 @@ public class Damageable : MonoBehaviour
             isAlive = value;
             animator.SetBool(AnimationStrings.isAlive, value);
             Debug.Log("IsAlive set " + value);
+
+            if (value == false)
+            {
+                damageableDeath.Invoke();
+            }
         }
     }
