+++ Assets/Scripts/FadeRemoveBehaviour.cs
@@ -5,7 +5,9 @@ using UnityEngine;
 public class FadeRemoveBehaviour : StateMachineBehaviour
 {
     public float fadeTime = 0.5f;
+    public float fadeDelay = 0.0f;
     private float timeElapsed = 0f;
+    private float fadeDelayElapsed = 0f;
     SpriteRenderer spriteRenderer;
     GameObject objToRemove;
     Color startColor;
@@ -22,15 +24,22 @@ public class FadeRemoveBehaviour : StateMachineBehaviour
     // OnStateUpdate is called on each Update frame between OnStateEnter and OnStateExit callbacks
     override public void OnStateUpdate(Animator animator, AnimatorStateInfo stateInfo, int layerIndex)
     {
-        timeElapsed += Time.deltaTime;
-
-        float newAlpha = startColor.a * (1 - (timeElapsed / fadeTime));
-
-        spriteRenderer.color = new Color(startColor.r, startColor.g, startColor.b, newAlpha);
-
-        if (timeElapsed > fadeTime)
+        if (fadeDelay > fadeDelayElapsed)
         {
-            Destroy(objToRemove);
+            fadeDelayElapsed += Time.deltaTime;
+        }
+        else
+        {
+            timeElapsed += Time.deltaTime;
+
+            float newAlpha = startColor.a * (1 - (timeElapsed / fadeTime));
+
+            spriteRenderer.color = new Color(startColor.r, startColor.g, startColor.b, newAlpha);
+
+            if (timeElapsed > fadeTime)
+            {
+                Destroy(objToRemove);
+            }
         }
     }
 }
