+++ Builds/PC/Prototype Platformer_Data/boot.config
@@ -3,4 +3,4 @@ gfx-enable-native-gfx-jobs=1
 wait-for-native-debugger=0
 hdr-display-enabled=0
 gc-max-time-slice=3
-build-guid=70277c9995254fd8997b7218c0cc0f0a
+build-guid=7e16171d443b4c46907a2580600c57cd
